# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EBAC",
        "description": "Documentation OpenAPI des API EBAC : églises, paramètres académiques et pré-inscription.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur courant"
        },
        {
            "url": "https://api-ebac.severinzran.ci/api/v1",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/parametres/matieres": {
            "get": {
                "tags": [
                    "Matières"
                ],
                "operationId": "listerMatieres",
                "parameters": [
                    {
                        "name": "niveau",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "obligatoire",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste filtrée des matières"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Matières"
                ],
                "operationId": "creerMatiere",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatierePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Matière créée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/parametres/matieres/{id}": {
            "get": {
                "tags": [
                    "Matières"
                ],
                "operationId": "afficherMatiere",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matière trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Matières"
                ],
                "operationId": "remplacerMatiere",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatierePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Matière modifiée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Matières"
                ],
                "operationId": "supprimerMatiere",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matière supprimée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Matières"
                ],
                "operationId": "modifierMatiere",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MatierePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Matière modifiée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MatierePayload": {
                "required": [
                    "code",
                    "libelle",
                    "id_niveau"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "MAT-BIB-001"
                    },
                    "libelle": {
                        "type": "string",
                        "example": "Introduction biblique"
                    },
                    "id_niveau": {
                        "type": "integer"
                    },
                    "coefficient": {
                        "type": "number",
                        "format": "float",
                        "default": 1
                    },
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "objectifs": {
                        "type": "string",
                        "nullable": true
                    },
                    "prerequis": {
                        "type": "string",
                        "nullable": true
                    },
                    "note_validation": {
                        "type": "number",
                        "format": "float",
                        "default": 10
                    },
                    "obligatoire": {
                        "type": "boolean",
                        "default": true
                    },
                    "active": {
                        "type": "boolean",
                        "default": true
                    },
                    "version": {
                        "type": "integer",
                        "default": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Jeton retourné après la connexion et la confirmation OTP.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}